Sync from dev branch to main

+++ scripts/api.py
@@ -557,6 +557,21 @@ def sagemaker_api(_, app: FastAPI):
     def ping():
         return {'status': 'Healthy'}
 
+import hashlib
+def md5(fname):
+    hash_md5 = hashlib.md5()
+    with open(fname, "rb") as f:
+        for chunk in iter(lambda: f.read(4096), b""):
+            hash_md5.update(chunk)
+    return hash_md5.hexdigest()
+
+def get_file_md5_dict(path):
+    file_dict = {}
+    for root, dirs, files in os.walk(path):
+        for file in files:
+            file_dict[file] = md5(os.path.join(root, file))
+    return file_dict
+
 def move_model_to_tmp(_, app: FastAPI):
     # os.system("rm -rf models")
     # Create model dir
@@ -566,13 +581,27 @@ def move_model_to_tmp(_, app: FastAPI):
     logging.info("Copy model dir to tmp")
     model_tmp_dir = f"models_{time.time()}"
     os.system(f"cp -rL models /tmp/{model_tmp_dir}")
-    os.system(f"rm -rf models")
-    # Delete tmp model dir
-    # print("Delete tmp model dir")
-    # os.system("rm -rf /tmp/models")
-    # Link model dir
-    logging.info("Link model dir")
-    os.system(f"ln -s /tmp/{model_tmp_dir} models")
+    src_file_dict = get_file_md5_dict("models")
+    tgt_file_dict = get_file_md5_dict(f"/tmp/{model_tmp_dir}")
+    is_complete = True
+    for file in src_file_dict:
+        logging.info(f"Src file {file} md5 {src_file_dict[md5]}")
+        if file not in tgt_file_dict:
+            is_complete = False
+            break
+        if src_file_dict[file] != tgt_file_dict[file]:
+            is_complete = False
+            break
+    if is_complete:
+        os.system(f"rm -rf models")
+        # Delete tmp model dir
+        # print("Delete tmp model dir")
+        # os.system("rm -rf /tmp/models")
+        # Link model dir
+        logging.info("Link model dir")
+        os.system(f"ln -s /tmp/{model_tmp_dir} models")
+    else:
+        logging.info("Failed to copy model dir, use the original dir")
     logging.info("Check disk usage on app started")
     os.system("df -h")
